+++ dgraph/cmd/bulk/reduce.go
@@ -39,6 +39,7 @@ import (
 	"github.com/dgraph-io/dgraph/protos/pb"
 	"github.com/dgraph-io/dgraph/worker"
 	"github.com/dgraph-io/dgraph/x"
+	"github.com/dgraph-io/ristretto/z"
 	"github.com/gogo/protobuf/proto"
 )
 
@@ -101,13 +102,22 @@ func (r *reducer) createBadger(i int) *badger.DB {
 		}
 	}
 
-	opt := badger.DefaultOptions(r.opt.shardOutputDirs[i]).WithSyncWrites(false).
-		WithTableLoadingMode(bo.MemoryMap).WithValueThreshold(1 << 10 /* 1 KB */).
-		WithLogger(nil).WithBlockCacheSize(1 << 20).
-		WithEncryptionKey(enc.ReadEncryptionKeyFile(r.opt.BadgerKeyFile))
+	opt := badger.DefaultOptions(r.opt.shardOutputDirs[i]).
+		WithSyncWrites(false).
+		WithTableLoadingMode(bo.MemoryMap).
+		WithValueThreshold(1 << 10 /* 1 KB */).
+		WithLogger(nil).
+		WithEncryptionKey(enc.ReadEncryptionKeyFile(r.opt.BadgerKeyFile)).
+		WithBlockCacheSize(r.opt.BlockCacheSize).
+		WithIndexCacheSize(r.opt.IndexCacheSize)
 
-	// TOOD(Ibrahim): Remove this once badger is updated.
-	opt.ZSTDCompressionLevel = 1
+	opt.Compression = bo.None
+	opt.ZSTDCompressionLevel = 0
+	// Overwrite badger options based on the options provided by the user.
+	if r.opt.BadgerCompressionLevel > 0 {
+		opt.Compression = bo.ZSTD
+		opt.ZSTDCompressionLevel = r.state.opt.BadgerCompressionLevel
+	}
 
 	db, err := badger.OpenManaged(opt)
 	x.Check(err)
@@ -162,7 +172,7 @@ func newMapIterator(filename string) *mapIterator {
 }
 
 func (r *reducer) encodeAndWrite(
-	writer *badger.StreamWriter, entryCh chan []*pb.MapEntry, closer *y.Closer) {
+	writer *badger.StreamWriter, entryCh chan []*pb.MapEntry, closer *z.Closer) {
 	defer closer.Done()
 
 	var listSize int
@@ -229,7 +239,7 @@ func (r *reducer) encodeAndWrite(
 
 func (r *reducer) reduce(mapItrs []*mapIterator, ci *countIndexer) {
 	entryCh := make(chan []*pb.MapEntry, 100)
-	closer := y.NewCloser(1)
+	closer := z.NewCloser(1)
 	defer closer.SignalAndWait()
 
 	var ph postingHeap